Fix executing commands as the controlling user (#67)
This patch changes how InGameCommandManager handles commands to be more like
ServerChatSpectatorHandler, to only take the first word into account when
checking which command to invoke. Previously, the switch statement for handling
the controller-only commands compared with the entire message, instead of only
the first word. This creates a problem for the "m" command, which expects an
argument.