Remap controller's tab entry UUID to forwarded
This patch remaps the player list entry for the controlling user, changing the
UUID from the logged-in account to the forwarded UUID of the client. This is
needed as the connecting Minecraft client receives its
"ClientboundGameProfilePacket" from the proxy it is connected to.

This is not needed to play, but without it the user's own skin will be shown as
one of the Minecraft defaults.