Remap UUID collisions for all users
Expands player list UUID rewriting to all users, including spectators. If a
player logs in and a user with the same UUID is already online on the server,
the UUID of the existing player needs to be remapped when sent to the client as
the client is logged in as their real UUID from the proxy. This is not an issue
without a proxy because ZenithProxy can send a custom GameProfilePacket when
logging in so the client is logged in as a different account.

This is a little bit of a messy system, will look into improving.